requirements.txt
# Core
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0

# Alpamayo
einops
hydra-core
omegaconf

# Vision
opencv-python
pillow
qwen-vl-utils

# Training (optional)
peft>=0.10.0
datasets
tqdm

# Server (optional)
fastapi
uvicorn
roslibpy
